backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.1
django-environ==0.12.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
elevenlabs==2.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
psycopg2==2.9.10
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.9.0
requests==2.32.3
sentry-sdk==2.29.1
sniffio==1.3.1
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
websockets==15.0.1
